tests: Move `exampleData.js` into the new `__tests__/lib` directory.
... which is intended for utility code of this sort. Adjust all its
imports accordingly (or let VSCode do so, at least).

This is technically a functional change, since the dummy test in
`exampleData.js` will no longer be run. (See `jest.config.js`.)
