+++ src/api/__tests__/queueMarkAsRead-test.js
@@ -1,7 +1,7 @@
 /* @flow strict-local */
 import queueMarkAsRead, { resetAll } from '../queueMarkAsRead';
 import * as messagesFlags from '../messages/messagesFlags';
-import * as eg from '../../__tests__/exampleData';
+import * as eg from '../../__tests__/lib/exampleData';
 import { Lolex } from '../../__tests__/lib/lolex';
 
 // $FlowFixMe Make flow understand about mocking
