+++ src/session/__tests__/sessionReducer-test.js
@@ -17,7 +17,7 @@ import {
   INITIAL_FETCH_START,
 } from '../../actionConstants';
 import sessionReducer from '../sessionReducer';
-import * as eg from '../../__tests__/exampleData';
+import * as eg from '../../__tests__/lib/exampleData';
 import { privateNarrow } from '../../utils/narrow';
 
 describe('sessionReducer', () => {
